Set addresses with cidr mask included
Currently the wireguard interfaces managed by the agent don't get
a bind address with a mask, it therefore defaults to /32 and no
routes are added in the table, which effectively breaks a lot of
networking b/w the hub and the spokes.

This isn't a super great solution but I was trying to isolate some
of the code-that-calls-Neutron-core to just the driver plugin API
and also trying to limit where _that_ is called. Which is during
port create/update. So, it will cache subnets it sees on updated
ports at this time, but that info isn't used until the main agent
loop runs.

This introduces a possible failure mode where the agent is restarted
and does not have all the info it needs when it starts up. To guard
against this, addresses on the wg interface will not be flushed
if this happens, to avoid removing good addresses. It's also not
something that will happen often, that a hub port's address would
actually change.