create hub port if missing on startup
The Wireguard agent takes action in two cases:
1. The periodic sync loop fires
2. The agent receives a RPC from Neutron

On each sync, the agent compares the "previous" to the "new" state from neutron, and acts to update the local interfaces to match.

However, we did not check for interfaces that changed "outside" of this process. In particular, previously created interfaces that are now missing.

This commit adds code to check for missing interfaces on startup, and create them from the saved configuration files.

TODO: We still don't check the "state" of the interface, only the difference between the neutron config and the wireguard config file. We need to do the following:

* Check the current state of all configured interfaces, and update them to the latest neutron state if needed.
* update the saved config file with the neutron state
* refactor to remove dependence on the saved config file, only the private key is not stored in the neutron DB.