wg hub is now scoped using a conf option
The previous approach to deciding whether to scope the wireguard hub
port to a project network namespace or the root network namespace relied
on a complicated boolean expression that was evaluated using options
from the neutron database, overall, creating too many scenarios in which
the scoping did not behave correctly. This change is meant to replace
the complex and esoteric expression with a simple configuration option
that decides the scoping. By default the wg hub port created by the
admin project in an effort to spin up a hub and spoke platform should be
scoped to the root network namespace.

This change however remains a temporary solution to a broader problem in
which hub ports should be automatically scoped depending on the project
by which they are created.